+++ lib/src/expression_analyzers/expression_initializer0.dart
@@ -109,6 +109,11 @@ class ExpressionInitializer0 extends ExpressionVisitor<void> {
   }
 
   @override
+  void visitStringChars(StringCharsExpression node) {
+    _initializeNode(node);
+  }
+
+  @override
   void visitSymbol(SymbolExpression node) {
     _initializeSymbol(node);
   }