+++ lib/src/expression_generators/expression_generators.dart
@@ -13,6 +13,7 @@ export 'ordered_choice_generator.dart';
 export 'repetition_generator.dart';
 export 'sequence_generator.dart';
 export 'slice_generator.dart';
+export 'string_char_generator.dart';
 export 'symbol_generator.dart';
 export 'verify_generator.dart';
 export 'zero_or_more_generator.dart';