+++ lib/src/expressions/expressions.dart
@@ -14,6 +14,7 @@ export 'repetition_expression.dart';
 export 'sep_by_expression.dart';
 export 'sequence_expression.dart';
 export 'slice_expression.dart';
+export 'string_chars_expression.dart';
 export 'symbol_expression.dart';
 export 'verify_expression.dart';
 export 'zero_or_more_expression.dart';